+++ src/server/http/ContentType.js
@@ -1,4 +1,13 @@
+/**
+ * Represents the Content-Type header.
+ */
 export default class ContentType {
+  /**
+   * The type, subtype, and charset of the content type,
+   * expressed in the HTTP header as follows:
+   *
+   *     [type]/[subtype]; charset=[charset]
+   */
   constructor (type, subtype, charset) {
     this.type = type
     this.subtype = subtype
@@ -14,6 +23,7 @@ export default class ContentType {
   }
 }
 
+// Some predefined content types
 ContentType.JSON = new ContentType('application', 'json', 'utf-8')
 ContentType.TEXT = new ContentType('text', 'plain', 'utf-8')
 ContentType.HTML = new ContentType('text', 'html', 'utf-8')