+++ src/server/http/Headers.js
@@ -1,3 +1,6 @@
+/**
+ * Represents HTTP headers.
+ */
 export default class Headers {
   constructor (raw = []) {
     this._raw = raw
@@ -5,12 +8,34 @@ export default class Headers {
     this._formatName = this._formatName.bind(this)
   }
 
+  /**
+   * Sets a header. The same header can have multiple values,
+   * in which case they will be concatinated with a comma,
+   * as described in the HTTP 1.1 specification.
+   *
+   * If you want to replace a header, you must clear it first.
+   *
+   *     headers
+   *       .clear('Origin')
+   *       .set('Origin', 'totally-real.net')
+   *
+   * @param {String} name
+   * @param value - Can be anything, but at the end, the #toString
+   *                method of this object dictates the actual header value.
+   * @returns {Headers}
+   */
   set (name, value) {
     return new Headers(
       this._raw.concat([[name.toLowerCase(), value]])
     )
   }
 
+  /**
+   * Gets a header, or an empty string.
+   *
+   * @param {String} name
+   * @returns {String}
+   */
   get (name) {
     const headerName = name.toLowerCase()
     return this._raw
@@ -20,10 +45,33 @@ export default class Headers {
       .join(', ')
   }
 
+  /**
+   * Clears a header.
+   */
+  clear (name) {
+    const headerName = name.toLowerCase()
+    return new Headers(
+      this._raw.filter(([name]) => name !== headerName)
+    )
+  }
+
+  /**
+   * Runs a function for every header.
+   *
+   * @param {(String, String) -> Void}
+   */
   forEach (callback) {
     this.map(callback)
   }
 
+  /**
+   * Runs a function for every header,
+   * returning an array containing the
+   * result of applying the function to every header.
+   *
+   * @param {(String, String) -> T}
+   * @returns {[T]}
+   */
   map (transform) {
     return this._raw
       .map(([name]) => name)
@@ -47,12 +95,5 @@ export default class Headers {
 
   inspect () {
     return 'Headers {\n  ' + this.toString().replace(/\n/g, '\n  ') + '\n}'
-  }
-
-  clear (name) {
-    const headerName = name.toLowerCase()
-    return new Headers(
-      this._raw.filter(([name]) => name !== headerName)
-    )
   }
 }