+++ src/server/http/HttpMessage.js
@@ -1,15 +1,23 @@
 import Headers from './Headers'
 
+/**
+ * Common functionality for requests and responses.
+ * Contains the headers as well as an immutable
+ * key/value store.
+ */
 export default class HttpMessage {
   constructor (headers) {
     this.headers = headers || new Headers()
     this._store = {}
   }
 
-  get _protected () {
-    return ['headers']
-  }
-
+  /**
+   * Sets a value.
+   *
+   * @param {String} key
+   * @param value
+   * @returns {HttpMessage} - Of same type as this
+   */
   set (key, value) {
     if (this._protected.includes(key)) {
       throw new Error(`Cannot set protected property ${key}`)
@@ -22,6 +30,12 @@ export default class HttpMessage {
     return copy
   }
 
+  /**
+   * Unsets a value.
+   *
+   * @param {String} key
+   * @returns {HttpMessage} - Of same type as this
+   */
   unset (key) {
     const copy = this._copy()
     Object.assign(copy, this._store)
@@ -32,5 +46,9 @@ export default class HttpMessage {
 
   _copy () {
     throw new Error('Implement this method')
+  }
+
+  get _protected () {
+    return ['headers']
   }
 }