+++ src/server/http/requestBody/JSONRequestBodyParser.js
@@ -1,0 +1,12 @@
+import parse from 'co-body'
+
+export default class JSONRequestBodyParser {
+  parse (stream) {
+    if (!stream.headers) {
+      stream.headers = {
+        'content-length': null
+      }
+    }
+    return parse.json(stream)
+  }
+}