+++ src/server/http/responseDecorators/jsxDecorator.js
@@ -2,6 +2,10 @@ import { isValidElement } from 'react'
 import { renderToStaticMarkup } from 'react-dom/server'
 import Response from '../Response'
 
+/**
+ * If a middleware returns JSX (a React element) this
+ * decorator turns it into an HTML response.
+ */
 export default function jsxDecorator (response) {
   if (isValidElement(response)) {
     return Response.ok(render(response))
@@ -13,6 +17,11 @@ function render (element) {
   return '<!DOCTYPE html>' + renderToStaticMarkup(enforceTrunk(element))
 }
 
+/**
+ * Whatever the root element's tag name is, this
+ * function normalizes it and makes sure there is
+ * always a proper <html> element and <body> in place.
+ */
 function enforceTrunk (element) {
   switch (element.type) {
     case 'html':